将数据处理成 barChart 能接受的形式。

+++ app/src/main/java/com/unircorn/csp/data/model/MediaPlaySummaryResponse.kt
@@ -3,5 +3,5 @@ package com.unircorn.csp.data.model
 data class MediaPlaySummaryResponse(
     val category: List<String>,
     val date: List<String>,
-    val value: List<List<Int>>
+    val value: List<List<Float>>
 )
